// web/src/api/generated/models/user-shelf-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface UserShelfDTO {
    'userId'?: number;
    'bookId'?: number;
    'bookTitle'?: string;
    'bookAuthor'?: string;
    'bookCoverUrl'?: string;
    'lastChapterIndex'?: number;
    'lastPosition'?: number;
    'readingProgress'?: number;
    'addedTime'?: string;
    'lastReadTime'?: string;
}